requirements.txt
fastapi==0.87.0
pydantic==1.10.2
uvicorn==0.19.0
python-jose==3.3.0
